% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOut.R
\name{customer_sync_byName}
\alias{customer_sync_byName}
\title{1.同步历史客户}
\usage{
customer_sync_byName(dmsToken, erpToken, customerName)
}
\arguments{
\item{dmsToken}{}

\item{erpToken}{}

\item{customerName}{}

\item{debug}{}
}
\value{
返回值
}
\description{
1.同步历史客户
}
\examples{
customer_sync_byName('CA64752E-79BB-4C59-8D7B-6C5849FC43F8','81F7F79A-7D3D-4114-A012-62888F8A6113','rdsTest2')
}
